# expression: implement vectorized evaluation for `builtinAesDecryptIVSig`